requirements.txt
streamlit>=1.30.0
openai>=1.10.0
pandas>=2.0.0
python-dotenv>=1.0.0
